Extract theme-builder sidebar controls into ag-website-shared
Moves the theme builder's left-sidebar UI primitives - color picker,
length input, collapsible accordion section, and searchable param
selector - into ag-website-shared/src/components/theme-builder, so
they can be shared with the AG Studio theme builder in a later PR.

- Input/FormattedInput: moved as-is (already host-agnostic)
- LengthInput: grid's CSS-length editor, with optional min/max
  clamping added as an opt-in (grid's own default behaviour is
  unchanged)
- ColorPicker: grid's hex input + swatch + react-colorful popover
- CollapsibleSection: accordion header + animated body, built on the
  existing shared Collapsible instead of Radix Accordion
- ParamSearchSelector: generalised from grid's AdvancedParamSelector
  via getKey/getLabel/getDocs/isEnabled/onToggle/renderEnabledItem
  props instead of being hardcoded to ParamModel

Grid's ColorValueEditor, LengthValueEditor, ScaleValueEditor,
AdvancedParamSelector, EditorPanel and GridPreview are rewired onto
these. Originals not touched by this pass (Input.tsx/FormattedInput.tsx
still used by CssValueEditor) are left in place.
