+++ documentation/ag-grid-docs/src/components/theme-builder/components/editors/LengthValueEditor.tsx
@@ -1,7 +1,6 @@
+import { LengthInput } from '@ag-website-shared/components/theme-builder/LengthInput';
 import { type LengthValue, paramValueToCss } from '@ag-website-shared/theming/api';
-import { reinterpretCSSValue, stripFloatingPointErrors } from '@ag-website-shared/theming/utils';
 
-import { FormattedInput } from './FormattedInput';
 import { type ValueEditorProps } from './ValueEditorProps';
 
 export const LengthValueEditor = ({
@@ -9,58 +8,10 @@ export const LengthValueEditor = ({
     value,
     onChange,
     icon,
-    swipeAdjustmentDivisor = 100,
+    swipeAdjustmentDivisor,
 }: ValueEditorProps<LengthValue>) => {
     const cssValue = paramValueToCss(param.property, value, null) || '';
-    const units = getUnit(cssValue);
     return (
-        <FormattedInput
-            value={cssValue}
-            onChange={onChange}
-            onClear={() => onChange(null)}
-            valueToDisplayString={toDisplayString}
-            valueToEditingString={toEditingString}
-            validateEditingString={(editingString) => {
-                const parsed = parseFloat(editingString);
-                return isNaN(parsed) ? null : `${parsed}${units}`;
-            }}
-            icon={icon}
-            getIconSwipeAdjustment={(value, pixels) => {
-                const proportion = parseFloat(value);
-                if (isNaN(proportion)) {
-                    return value;
-                }
-
-                const rawAdjustment = parseFloat(Math.max(proportion + pixels / swipeAdjustmentDivisor, 0).toFixed(1));
-                return stripFloatingPointErrors(rawAdjustment) + units;
-            }}
-        />
+        <LengthInput value={cssValue} onChange={onChange} icon={icon} swipeAdjustmentDivisor={swipeAdjustmentDivisor} />
     );
 };
-
-const toDisplayString = (value: string) => {
-    value = value.trim();
-    const reinterpreted = reinterpretCSSValue(value, 'length') || value;
-    const unit = getUnit(reinterpreted);
-    const parsed = parseFloat(reinterpreted);
-    return isNaN(parsed) ? value : round2dp(parsed) + unit;
-};
-
-const toEditingString = (value: string): string => {
-    value = value.trim();
-    const reinterpreted = reinterpretCSSValue(value, 'length') || value;
-    const number = parseFloat(reinterpreted);
-    return isNaN(number) ? value : round2dp(number);
-};
-
-const cssFunctionRegex = /\w+\(/i;
-
-const getUnit = (value: string) => {
-    if (cssFunctionRegex.test(value)) {
-        value = reinterpretCSSValue(value, 'length') || value;
-    }
-    const [unit] = value.match(/[^\d.]+$/) || [];
-    return unit || '';
-};
-
-const round2dp = (value: number) => stripFloatingPointErrors(Math.round(value * 100) / 100);
