+++ documentation/ag-grid-docs/src/components/theme-builder/components/editors/ScaleValueEditor.tsx
@@ -1,8 +1,8 @@
+import { FormattedInput } from '@ag-website-shared/components/theme-builder/FormattedInput';
 import { stripFloatingPointErrors } from '@ag-website-shared/theming/utils';
 
 import type { ScaleValue } from 'ag-grid-community';
 
-import { FormattedInput } from './FormattedInput';
 import { type ValueEditorProps } from './ValueEditorProps';
 
 export const ScaleValueEditor = ({
