+++ external/ag-website-shared/src/components/theme-builder/formatted-input-global-styles.css
@@ -1,0 +1,3 @@
+.force-resize-cursor * {
+    cursor: ew-resize !important;
+}
